Move image optimization to plugin
