+++ packages/open-next/src/adapters/image-optimization-adapter.ts
@@ -1,9 +1,5 @@
-import { IncomingMessage, ServerResponse } from "node:http";
-import https from "node:https";
 import path from "node:path";
-import { Writable } from "node:stream";
 
-import { GetObjectCommand, S3Client } from "@aws-sdk/client-s3";
 import type {
   APIGatewayProxyEvent,
   APIGatewayProxyEventHeaders,
@@ -14,21 +10,18 @@ import type {
 // @ts-ignore
 import { defaultConfig } from "next/dist/server/config-shared";
 import {
-  imageOptimizer,
   ImageOptimizerCache,
   // @ts-ignore
 } from "next/dist/server/image-optimizer";
-// @ts-ignore
-import type { NextUrlWithParsedQuery } from "next/dist/server/request-meta";
 
+// @ts-ignore
 import { loadConfig } from "./config/util.js";
-import { awsLogger, debug, error } from "./logger.js";
+import { debug } from "./logger.js";
+import { optimizeImage } from "./plugins/image-optimization.js";
 import { setNodeEnv } from "./util.js";
 
 // Expected environment variables
-const { BUCKET_NAME, BUCKET_KEY_PREFIX } = process.env;
-
-const s3Client = new S3Client({ logger: awsLogger });
+const { BUCKET_NAME } = process.env;
 
 setNodeEnv();
 const nextDir = path.join(__dirname, ".next");
@@ -64,7 +57,7 @@ export async function handler(
       headers,
       queryString === null ? undefined : queryString,
     );
-    const result = await optimizeImage(headers, imageParams);
+    const result = await optimizeImage(headers, imageParams, nextConfig);
 
     return buildSuccessResponse(result);
   } catch (e: any) {
@@ -110,23 +103,6 @@ function validateImageParams(
   return imageParams;
 }
 
-async function optimizeImage(
-  headers: APIGatewayProxyEventHeaders,
-  imageParams: any,
-) {
-  const result = await imageOptimizer(
-    // @ts-ignore
-    { headers },
-    {}, // res object is not necessary as it's not actually used.
-    imageParams,
-    nextConfig,
-    false, // not in dev mode
-    downloadHandler,
-  );
-  debug("optimized result", result);
-  return result;
-}
-
 function buildSuccessResponse(result: any) {
   return {
     statusCode: 200,
@@ -152,68 +128,4 @@ function buildFailureResponse(e: any) {
     },
     body: e?.message || e?.toString() || e,
   };
-}
-
-async function downloadHandler(
-  _req: IncomingMessage,
-  res: ServerResponse,
-  url: NextUrlWithParsedQuery,
-) {
-  // downloadHandler is called by Next.js. We don't call this function
-  // directly.
-  debug("downloadHandler url", url);
-
-  // Reads the output from the Writable and writes to the response
-  const pipeRes = (w: Writable, res: ServerResponse) => {
-    w.pipe(res)
-      .once("close", () => {
-        res.statusCode = 200;
-        res.end();
-      })
-      .once("error", (err) => {
-        error("Failed to get image", err);
-        res.statusCode = 400;
-        res.end();
-      });
-  };
-
-  try {
-    // Case 1: remote image URL => download the image from the URL
-    if (url.href.toLowerCase().match(/^https?:\/\//)) {
-      pipeRes(https.get(url), res);
-    }
-    // Case 2: local image => download the image from S3
-    else {
-      // Download image from S3
-      // note: S3 expects keys without leading `/`
-      const keyPrefix = BUCKET_KEY_PREFIX?.replace(/^\/|\/$/g, "");
-      const response = await s3Client.send(
-        new GetObjectCommand({
-          Bucket: BUCKET_NAME,
-          Key: keyPrefix
-            ? keyPrefix + "/" + url.href.replace(/^\//, "")
-            : url.href.replace(/^\//, ""),
-        }),
-      );
-
-      if (!response.Body) {
-        throw new Error("Empty response body from the S3 request.");
-      }
-
-      // @ts-ignore
-      pipeRes(response.Body, res);
-
-      // Respect the bucket file's content-type and cache-control
-      // imageOptimizer will use this to set the results.maxAge
-      if (response.ContentType) {
-        res.setHeader("Content-Type", response.ContentType);
-      }
-      if (response.CacheControl) {
-        res.setHeader("Cache-Control", response.CacheControl);
-      }
-    }
-  } catch (e: any) {
-    error("Failed to download image", e);
-    throw e;
-  }
 }
